Fix implicit conversion

+++ plugin/source/PluginProcessor.cpp
@@ -198,7 +198,7 @@ void AudioPluginAudioProcessor::processBlock(juce::AudioBuffer<float>& buffer,
 
   const auto inBlock =
       juce::dsp::AudioBlock<float>{buffer}.getSubsetChannelBlock(
-          0u, getTotalNumOutputChannels());
+          0u, static_cast<size_t>(getTotalNumOutputChannels()));
   auto outBlock = juce::dsp::AudioBlock<float>{envelopeFollowerOutputBuffer};
   envelopeFollower.process(
       juce::dsp::ProcessContextNonReplacing<float>{inBlock, outBlock});
